Update to Swagger UI v3.10.0
This commit updates the Swagger UI code in the dist folder to tag
v3.10.0.  The files in `dist` were replaced and `index.html` was
renamed to `_index.html`.  The latest Swagger UI index was used as
basis for `templates/index.html`.

OAuth2 configuration was based on: swagger-api/swagger-ui/docs/usage/oauth2.md